Module Tar_cstruct

Processing tar content with cstruct buffers

type in_channel

type out_channel

val make_in_channel : Cstruct.t -> in_channel

make_in_channel buf uses buf as a source of raw tar content.

val make_out_channel : unit -> out_channel

make_out_channel () returns a buffer to hold serialized tar content.

val to_string : out_channel -> string

to_string oc returns the contents of oc as a string of bytes.

val to_cstruct : out_channel -> Cstruct.t

to_cstruct oc returns the contents of oc as a Cstruct.t.

val really_read : in_channel -> Cstruct.t -> unit

really_read ic buf fills buf with data from ic or raises Pervasives.End_of_file

val really_write : out_channel -> Cstruct.t -> unit

really_write oc buf writes the full contents of buf to oc or raises End_of_file

module Header : sig ... end

module Archive : sig ... end