Module Tar_gz

type error = [

| `Fatal of Tar.error

| `Eof

| `Gz of string

]

val in_gzipped : ('a, [> error ] as 'err, 't) Tar.t -> ('a, 'err, 't) Tar.t

in_gzipped takes a tar process (like Tar.fold) and add a uncompression layer to be able to manipulate a *.tar.gz archive.

val out_gzipped : level:int -> mtime:int32 -> Gz.os -> ('a, 'err, 't) Tar.t -> ('a, 'err, 't) Tar.t

out_gzipped takes a tar process (like Tar.out) and add a compression layer to be able to generate a *.tar.gz archive.