Module Tar.Make

Parameters

module IO : IO

Signature

val really_read : IO.in_channel -> Cstruct.t -> unit

really_read fd buf fills buf with data from fd or raises End_of_file

val really_write : IO.out_channel -> Cstruct.t -> unit

really_write fd buf writes the full contents of buf to fd or raises End_of_file

module Header : sig ... end

val write_block : ?⁠level:Header.compatibility -> Header.t -> (IO.out_channel -> unit) -> IO.out_channel -> unit

Write hdr, then call write_body fd to write the body, then zero-pads so the stream is positioned for the next block.

val write_end : IO.out_channel -> unit

Writes a stream terminator to fd

module Archive : sig ... end