Module Tar_cstruct.Archive

val with_next_file : in_channel -> (in_channel -> Header.t -> 'a) -> 'a

with_next_file ic f Read the next header, apply the function f to ic and the header. The function should leave ic positioned immediately after the datablock. really_read can be used for this purpose. Finally the function skips past the zero padding to the next header

val list : ?⁠level:Header.compatibility -> in_channel -> Header.t list

List the contents of a tar